Added basic rewind
Added the basic rewind functionality, and an interface to interact with the object when clicking 'E'. It rewind both location and rotation, but it happens in a while loop so it looks like a teleport. Need to figure out how to make it lerp in UE5. I would use a coroutine or async await task in Unity.